floatpoint32 computetype

+++ backend/live-transcription/app/transcribe.py
@@ -2,7 +2,7 @@ from faster_whisper import WhisperModel
 import time
 
 model_size = "base"
-model = WhisperModel(model_size, device ="cuda", compute_type="int8")
+model = WhisperModel(model_size, device ="cuda", compute_type="float32")
 
 def transcribe(file_path: str) -> dict:
     print(f"Transcription of {file_path} started...")
